+++ tests/testthat/helper-for-tests.R
@@ -1,6 +1,4 @@
-# Function factories to create object only the first time they are accessed
-
-## Simulation function factory
+# Simulation function factory
 getTestSimulation <- function() {
   .simulation <- NULL
   function() {
@@ -17,18 +15,18 @@ testSimulations <- function() {
   list("Aciclovir" = testSimulation())
 }
 
-## observedData function factory
+# Observed data function factory
 getTestObservedData <- function() {
   .observedData <- NULL
   function() {
     if (is.null(.observedData)) {
       filePath <- testthat::test_path("../data/AciclovirLaskinData.xlsx")
-      dataConfiguration <- createImporterConfigurationForFile(filePath = filePath)
-      dataConfiguration$sheets <- "Laskin 1982.Group A"
-      dataConfiguration$namingPattern <- "{Source}.{Sheet}"
+      dataConfig <- createImporterConfigurationForFile(filePath)
+      dataConfig$sheets <- "Laskin 1982.Group A"
+      dataConfig$namingPattern <- "{Source}.{Sheet}"
       .observedData <<- loadDataSetsFromExcel(
         xlsFilePath = filePath,
-        importerConfigurationOrPath = dataConfiguration
+        importerConfigurationOrPath = dataConfig
       )
     }
     return(.observedData)
@@ -37,7 +35,7 @@ getTestObservedData <- function() {
 
 testObservedData <- getTestObservedData()
 
-## Parameters function factory
+# Parameters function factory
 getTestParameters <- function() {
   .parameters <- NULL
   function() {
@@ -66,40 +64,67 @@ getTestParameters <- function() {
 
 testParameters <- getTestParameters()
 
-
-## OutputMapping function factory
+# OutputMapping function factory
 getTestOutputMapping <- function(includeObservedData = TRUE) {
   .outputMapping <- NULL
-
   function() {
     if (is.null(.outputMapping)) {
-      outputMapping <- PIOutputMapping$new(
+      mapping <- PIOutputMapping$new(
         quantity = getQuantity(
           "Organism|PeripheralVenousBlood|Aciclovir|Plasma (Peripheral Venous Blood)",
           container = testSimulations()$Aciclovir
         )
       )
-
       if (includeObservedData) {
-        outputMapping$addObservedDataSets(
+        mapping$addObservedDataSets(
           testObservedData()$`AciclovirLaskinData.Laskin 1982.Group A`
         )
       }
-
-      .outputMapping <<- list(outputMapping)
+      .outputMapping <<- list(mapping)
     }
     return(.outputMapping)
   }
 }
 
-# Example usage
 testOutputMapping <- getTestOutputMapping()
-testOutputMappingWoObservedData <- getTestOutputMapping(includeObservedData = FALSE)
+testOutputMappingWithoutObsData <- getTestOutputMapping(includeObservedData = FALSE)
 
+# Function to create a ParameterIdentification task
+createPiTask <- function() {
+  ParameterIdentification$new(
+    simulations = testSimulation(),
+    parameters = testParameters(),
+    outputMappings = testOutputMapping(),
+    configuration = NULL
+  )
+}
 
+# Function to create a modified ParameterIdentification task (simulation failure)
+getTestModifiedTask <- function() {
+  function() {
+    sim <- loadSimulation(system.file("extdata", "Aciclovir.pkml", package = "ospsuite"))
+    sim$solver$mxStep <- 1
+    mapping <- PIOutputMapping$new(
+      quantity = getQuantity("Organism|PeripheralVenousBlood|Aciclovir|Plasma (Peripheral Venous Blood)", container = sim)
+    )
+    mapping$addObservedDataSets(testObservedData())
 
-# PI multiple simulations and parameter paths
+    params <- PIParameters$new(parameters = list(
+      getParameter("Aciclovir|Lipophilicity", container = sim)
+    ))
 
+    task <- ParameterIdentification$new(
+      simulations = sim,
+      parameters = params,
+      outputMappings = mapping
+    )
+    return(task)
+  }
+}
+
+testModifiedTask <- getTestModifiedTask()
+
+# Multiple simulation and parameter paths
 sim_250mg <- loadSimulation(
   system.file("extdata", "Aciclovir.pkml", package = "ospsuite")
 )
@@ -110,6 +135,9 @@ sim_500mg <- loadSimulation(
 piParameterLipo <- PIParameters$new(parameters = list(
   getParameter(path = "Aciclovir|Lipophilicity", container = sim_250mg),
   getParameter(path = "Aciclovir|Lipophilicity", container = sim_500mg)
+))
+piParameterLipo_250mg <- PIParameters$new(parameters = list(
+  getParameter(path = "Aciclovir|Lipophilicity", container = sim_250mg)
 ))
 piParameterCl_250mg <- PIParameters$new(
   parameters = getParameter(
@@ -123,9 +151,6 @@ piParameterCl_500mg <- PIParameters$new(
     container = sim_500mg
   )
 )
-
-# outputMapping_250mg <- testOutputMapping()
-# outputMapping_500mg <- testOutputMapping()
 
 simOutputPath <- "Organism|PeripheralVenousBlood|Aciclovir|Plasma (Peripheral Venous Blood)"
 outputMapping_250mg <- PIOutputMapping$new(
@@ -141,8 +166,7 @@ outputMapping_500mg$addObservedDataSets(
   testObservedData()$`AciclovirLaskinData.Laskin 1982.Group A`
 )
 
-# Variables
-
+# Other variables
 testQuantity <- ospsuite::getQuantity(
   path = "Organism|PeripheralVenousBlood|Aciclovir|Plasma (Peripheral Venous Blood)",
   container = testSimulation()
@@ -154,7 +178,6 @@ piConfiguration <- PIConfiguration$new()
 
 
 # Helper functions
-
 getTestDataFilePath <- function(fileName) {
   dataPath <- testthat::test_path("../data")
   file.path(dataPath, fileName, fsep = .Platform$file.sep)
@@ -164,28 +187,16 @@ getSimulationFilePath <- function(simulationName) {
   getTestDataFilePath(paste0(simulationName, ".pkml"))
 }
 
-# Helper function to load a model easily. In the test environment, we do not
-# want to load from cache by default. Instead new instances should be created
-# unless specifically specified otherwise
 loadTestSimulation <- function(simulationName, loadFromCache = FALSE, addToCache = TRUE) {
   simFile <- getSimulationFilePath(simulationName)
-  sim <- ospsuite::loadSimulation(simFile, loadFromCache = loadFromCache, addToCache = addToCache)
+  sim <- ospsuite::loadSimulation(
+    simFile,
+    loadFromCache = loadFromCache, addToCache = addToCache
+  )
 }
 
 executeWithTestFile <- function(actionWithFile) {
   newFile <- tempfile()
   actionWithFile(newFile)
   file.remove(newFile)
-}
-
-# Helper function to create a parameter identification task
-# Start with clean configuration state to avoid side effects from shared configurations
-createPiTask <- function() {
-  piTask <- ParameterIdentification$new(
-    simulations = testSimulation(),
-    parameters = testParameters(),
-    outputMappings = testOutputMapping(),
-    configuration = NULL
-  )
-  return(piTask)
 }